Added Doctor model.

+++ sql/patients.sql
@@ -1,5 +1,5 @@
 create table if not exists patients(
-    patient_id serial primary key,
+    patient_id BIGINT primary key,
     first_name VARCHAR(20) not null,
     last_name VARCHAR(30) not null,
     patient_email VARCHAR(30) not null,
