+++ sql/users.sql
@@ -1,5 +1,5 @@
 create table if not exists users (
-    user_id BIGINT primary key,
+    user_id serial primary key,
     first_name VARCHAR(20) not null,
     last_name VARCHAR(30) not null,
     user_email VARCHAR(30) not null,
